Homework finish

+++ Korostelev_Mihail/home_assignment4/taskTwo/app/app.js
@@ -1,6 +1,6 @@
-for (var userNumber = Number.parseFloat(prompt("Enter your value (to close enter 0)"));
+for (var userNumber = Number.parseInt(prompt("Enter your value (to close enter 0)"));
   userNumber > 0;
-  userNumber = parseFloat(prompt("Enter your value (to close enter 0)"))
+  userNumber = parseInt(prompt("Enter your value (to close enter 0)"))
 ) {
   if (userNumber == 0) {
     break;
